Add controller & data-access layer for books operations
